Refactoring for multi-core support
Currently doesn't work due to missing memory atomic operations support.
Also needs interrupt rework.

This commit changes PLIC CPU interrupt mapping, see src/devices/plic.c

- Made possible to create multiple harts with same devices. Will need
  to get rid of 'global_vm_list' variable in the future and make
  something like 'rvvm_domain' to hold multiple separate VMs.
- wait_event is now atomic. Doesn't solve the interrupt problem
  because setting ev_int_mask requires read-modify-write operation
  and clearing it's value can lead to missing interrupts.
- Added a header for atomic operations - atomic.h. Relies on C11
  atomics.
- Put 'static inline' instead of 'inline' everywhere due to C inlining
  behavior. Makes -O0 build working.

+++ src/bit_ops.h
@@ -28,7 +28,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // Simple math operations (sign-extend int bits, etc) for internal usage
 // this entire header is cursed (though everything optimizes fine)
 
-#define LONG_BITS (CHAR_BIT * sizeof(long))
+#define LONG_BITS (CHAR_BIT * sizeof(ptrdiff_t))
 
 /*
 * Sign-extend bits in the lower part of val into signed long
@@ -37,7 +37,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 *
 *     [ext is now equal to signed lower 20 bits of val]
 */
-static inline long sign_extend(long val, uint8_t bits)
+static inline long sign_extend(ptrdiff_t val, uint8_t bits)
 {
     return (val << (LONG_BITS - bits)) >> (LONG_BITS - bits);
 }
